# Improve Metal runtime command buffer handling