This PR changes how we handle command buffers in the Metal runtime.

Previously, we waited for each command buffer to finish before queuing more work, which according to Apple's [Tuning Hints](https://developer.apple.com/documentation/metalperformanceshaders/tuning_hints) could introduce delays of up to 2.5ms. 

Now, we've implemented a dispatch group approach. We enter the group when we create a new command buffer and leave when it's done. 

This modification allows the `synchronize` method to wait on the group instead of individual buffers, which should enhance both the speed and stability of the Metal runtime.