@tinyb0t 2023-07-24
Changes made in `tinygrad/`:
```
------------------------------------------------------------
files                             insertions       deletions
------------------------------------------------------------
tinygrad/runtime/ops_metal.py             10               4
------------------------------------------------------------
lines added in the tinygrad folder: 6
```