@dc-dc-dc 2023-07-24
can you post before/after average test suite times with this change