@geohot 2023-07-24
I second this, can you post a benchmark? Overall I like the change