@zboralski 2023-07-25
> I second this, can you post a benchmark? Overall I like the change

The performance is generally comparable between the two versions. Benchmark results can fluctuate quite a bit across different runs. Is that something you'd like to improve? 


before
```
❯ METAL=1 python test/test_speed_v_torch.py
add                                1x    1    0.01 ms (    0.00 GFLOPS     0.00 GB/s) in torch,    0.34 ms (    0.00 GFLOPS     0.00 GB/s) in tinygrad,   48.71x slower       0.00 MOPS     0.00 MB
add                             1024x 1024    0.50 ms (    2.08 GFLOPS    24.98 GB/s) in torch,    0.71 ms (    1.48 GFLOPS    17.71 GB/s) in tinygrad,    1.41x slower       1.05 MOPS    12.58 MB
add                             4096x 4096    6.76 ms (    2.48 GFLOPS    29.77 GB/s) in torch,    3.90 ms (    4.30 GFLOPS    51.61 GB/s) in tinygrad,    0.58x faster      16.78 MOPS   201.33 MB
add_constant                    4096x 4096    5.63 ms (    2.98 GFLOPS    23.84 GB/s) in torch,    2.76 ms (    6.08 GFLOPS    48.60 GB/s) in tinygrad,    0.49x faster      16.78 MOPS   134.22 MB
add_sq                          4096x 4096   19.40 ms (    2.59 GFLOPS    10.38 GB/s) in torch,    3.86 ms (   13.04 GFLOPS    52.15 GB/s) in tinygrad,    0.20x faster      50.33 MOPS   201.33 MB
array_packing                   2048x 2048    2.34 ms (    1.79 GFLOPS    14.36 GB/s) in torch,    1.07 ms (    3.94 GFLOPS    31.49 GB/s) in tinygrad,    0.46x faster       4.19 MOPS    33.55 MB
conv bs: 32 chans:  4 ->  32 k:3              1.78 ms (   42.30 GFLOPS     2.68 GB/s) in torch,    0.52 ms (  144.44 GFLOPS     9.17 GB/s) in tinygrad,    0.29x faster      75.50 MOPS     4.79 MB
conv bs: 32 chans: 16 ->  32 k:3              2.96 ms (  102.04 GFLOPS     2.22 GB/s) in torch,    0.83 ms (  361.83 GFLOPS     7.88 GB/s) in tinygrad,    0.28x faster     301.99 MOPS     6.58 MB
conv bs: 32 chans: 64 ->  32 k:3             11.98 ms (  100.79 GFLOPS     1.15 GB/s) in torch,    2.06 ms (  585.13 GFLOPS     6.65 GB/s) in tinygrad,    0.17x faster    1207.96 MOPS    13.74 MB
double_permute (64, 64, 64, 64)              12.42 ms (    1.35 GFLOPS    10.80 GB/s) in torch,    5.01 ms (    3.35 GFLOPS    26.80 GB/s) in tinygrad,    0.40x faster      16.78 MOPS   134.22 MB
exp                             2048x 2048    9.87 ms (    0.85 GFLOPS     3.40 GB/s) in torch,    1.06 ms (    7.92 GFLOPS    31.70 GB/s) in tinygrad,    0.11x faster       8.39 MOPS    33.55 MB
gemm                            1024x 1024    2.21 ms (  973.38 GFLOPS     5.70 GB/s) in torch,    2.07 ms ( 1036.18 GFLOPS     6.07 GB/s) in tinygrad,    0.94x faster    2147.48 MOPS    12.58 MB
gemm                             256x  256    0.06 ms (  529.46 GFLOPS    12.41 GB/s) in torch,    0.41 ms (   81.99 GFLOPS     1.92 GB/s) in tinygrad,    6.46x slower      33.55 MOPS     0.79 MB
gemm_unrolled                    512x  512    0.39 ms (  688.52 GFLOPS     8.07 GB/s) in torch,    0.63 ms (  425.38 GFLOPS     4.98 GB/s) in tinygrad,    1.62x slower     268.44 MOPS     3.15 MB
gemm_unrolled_permute_l          512x  512    0.36 ms (  741.45 GFLOPS     8.69 GB/s) in torch,    0.64 ms (  417.61 GFLOPS     4.89 GB/s) in tinygrad,    1.78x slower     268.44 MOPS     3.15 MB
gemm_unrolled_permute_lr         512x  512    0.35 ms (  769.71 GFLOPS     9.02 GB/s) in torch,    0.60 ms (  443.73 GFLOPS     5.20 GB/s) in tinygrad,    1.73x slower     268.44 MOPS     3.15 MB
gemm_unrolled_permute_r          512x  512    0.35 ms (  768.33 GFLOPS     9.00 GB/s) in torch,    0.65 ms (  415.03 GFLOPS     4.86 GB/s) in tinygrad,    1.85x slower     268.44 MOPS     3.15 MB
max                             4096x 4096    1.52 ms (   11.03 GFLOPS    44.12 GB/s) in torch,    4.54 ms (    3.69 GFLOPS    14.77 GB/s) in tinygrad,    2.99x slower      16.78 MOPS    67.11 MB
mul_sum                         4096x 4096    8.68 ms (    3.87 GFLOPS    15.47 GB/s) in torch,    8.47 ms (    3.96 GFLOPS    15.84 GB/s) in tinygrad,    0.98x faster      33.55 MOPS   134.22 MB
neg                             4096x 4096    5.71 ms (    2.94 GFLOPS    23.53 GB/s) in torch,    2.76 ms (    6.09 GFLOPS    48.69 GB/s) in tinygrad,    0.48x faster      16.78 MOPS   134.22 MB
conv bs:  1 chans: 12 ->  32 k:3              1.30 ms (   43.71 GFLOPS     1.12 GB/s) in torch,    1.19 ms (   47.68 GFLOPS     1.23 GB/s) in tinygrad,    0.92x faster      56.62 MOPS     1.46 MB
partial_sum                     4096x 4096    1.35 ms (   12.43 GFLOPS    49.72 GB/s) in torch,    2.13 ms (    7.86 GFLOPS    31.45 GB/s) in tinygrad,    1.58x slower      16.78 MOPS    67.11 MB
permute                         1024x 1024    1.20 ms (    0.88 GFLOPS     7.01 GB/s) in torch,    0.60 ms (    1.76 GFLOPS    14.06 GB/s) in tinygrad,    0.50x faster       1.05 MOPS     8.39 MB
permute                         4096x 4096   22.92 ms (    0.73 GFLOPS     5.86 GB/s) in torch,    2.78 ms (    6.04 GFLOPS    48.29 GB/s) in tinygrad,    0.12x faster      16.78 MOPS   134.22 MB
pow                             2048x 2048   20.93 ms (    4.81 GFLOPS     2.41 GB/s) in torch,    1.39 ms (   72.60 GFLOPS    36.30 GB/s) in tinygrad,    0.07x faster     100.66 MOPS    50.33 MB
relu                            4096x 4096    5.62 ms (    2.99 GFLOPS    23.88 GB/s) in torch,    2.75 ms (    6.10 GFLOPS    48.84 GB/s) in tinygrad,    0.49x faster      16.78 MOPS   134.22 MB
sub                             4096x 4096    6.86 ms (    2.45 GFLOPS    29.37 GB/s) in torch,    3.89 ms (    4.32 GFLOPS    51.80 GB/s) in tinygrad,    0.57x faster      16.78 MOPS   201.33 MB
sum                             2048x 2048    0.49 ms (    8.48 GFLOPS    33.91 GB/s) in torch,    1.24 ms (    3.37 GFLOPS    13.48 GB/s) in tinygrad,    2.52x slower       4.19 MOPS    16.78 MB
sum                             4096x 4096    1.33 ms (   12.58 GFLOPS    50.30 GB/s) in torch,    4.53 ms (    3.70 GFLOPS    14.80 GB/s) in tinygrad,    3.40x slower      16.78 MOPS    67.11 MB
.
----------------------------------------------------------------------
Ran 31 tests in 6.636s

OK (skipped=8)

```

after
```
❯ METAL=1 python test/test_speed_v_torch.py
add                                1x    1    0.01 ms (    0.00 GFLOPS     0.00 GB/s) in torch,    0.36 ms (    0.00 GFLOPS     0.00 GB/s) in tinygrad,   46.52x slower       0.00 MOPS     0.00 MB
add                             1024x 1024    0.52 ms (    2.03 GFLOPS    24.34 GB/s) in torch,    0.76 ms (    1.39 GFLOPS    16.62 GB/s) in tinygrad,    1.46x slower       1.05 MOPS    12.58 MB

add                             4096x 4096    6.56 ms (    2.56 GFLOPS    30.71 GB/s) in torch,    3.93 ms (    4.27 GFLOPS    51.24 GB/s) in tinygrad,    0.60x faster      16.78 MOPS   201.33 MB
add_constant                    4096x 4096    5.50 ms (    3.05 GFLOPS    24.40 GB/s) in torch,    2.77 ms (    6.06 GFLOPS    48.46 GB/s) in tinygrad,    0.50x faster      16.78 MOPS   134.22 MB
add_sq                          4096x 4096   19.33 ms (    2.60 GFLOPS    10.42 GB/s) in torch,    3.92 ms (   12.84 GFLOPS    51.35 GB/s) in tinygrad,    0.20x faster      50.33 MOPS   201.33 MB
array_packing                   2048x 2048    2.46 ms (    1.71 GFLOPS    13.65 GB/s) in torch,    1.06 ms (    3.97 GFLOPS    31.75 GB/s) in tinygrad,    0.43x faster       4.19 MOPS    33.55 MB
conv bs: 32 chans:  4 ->  32 k:3              1.79 ms (   42.22 GFLOPS     2.68 GB/s) in torch,    0.70 ms (  107.55 GFLOPS     6.82 GB/s) in tinygrad,    0.39x faster      75.50 MOPS     4.79 MB
conv bs: 32 chans: 16 ->  32 k:3              2.95 ms (  102.51 GFLOPS     2.23 GB/s) in torch,    0.88 ms (  343.89 GFLOPS     7.49 GB/s) in tinygrad,    0.30x faster     301.99 MOPS     6.58 MB
conv bs: 32 chans: 64 ->  32 k:3             12.00 ms (  100.68 GFLOPS     1.15 GB/s) in torch,    2.12 ms (  570.43 GFLOPS     6.49 GB/s) in tinygrad,    0.18x faster    1207.96 MOPS    13.74 MB
double_permute (64, 64, 64, 64)              12.38 ms (    1.36 GFLOPS    10.84 GB/s) in torch,    5.06 ms (    3.32 GFLOPS    26.53 GB/s) in tinygrad,    0.41x faster      16.78 MOPS   134.22 MB
exp                             2048x 2048    9.80 ms (    0.86 GFLOPS     3.42 GB/s) in torch,    1.06 ms (    7.94 GFLOPS    31.75 GB/s) in tinygrad,    0.11x faster       8.39 MOPS    33.55 MB
gemm                            1024x 1024    2.38 ms (  902.78 GFLOPS     5.29 GB/s) in torch,    2.06 ms ( 1041.23 GFLOPS     6.10 GB/s) in tinygrad,    0.87x faster    2147.48 MOPS    12.58 MB
gemm                             256x  256    0.06 ms (  517.54 GFLOPS    12.13 GB/s) in torch,    0.57 ms (   59.37 GFLOPS     1.39 GB/s) in tinygrad,    8.72x slower      33.55 MOPS     0.79 MB
gemm_unrolled                    512x  512    0.39 ms (  690.81 GFLOPS     8.10 GB/s) in torch,    0.66 ms (  403.97 GFLOPS     4.73 GB/s) in tinygrad,    1.71x slower     268.44 MOPS     3.15 MB
gemm_unrolled_permute_l          512x  512    0.37 ms (  732.93 GFLOPS     8.59 GB/s) in torch,    0.67 ms (  401.37 GFLOPS     4.70 GB/s) in tinygrad,    1.83x slower     268.44 MOPS     3.15 MB
gemm_unrolled_permute_lr         512x  512    0.35 ms (  760.17 GFLOPS     8.91 GB/s) in torch,    0.65 ms (  415.56 GFLOPS     4.87 GB/s) in tinygrad,    1.83x slower     268.44 MOPS     3.15 MB
gemm_unrolled_permute_r          512x  512    0.35 ms (  767.05 GFLOPS     8.99 GB/s) in torch,    0.67 ms (  398.22 GFLOPS     4.67 GB/s) in tinygrad,    1.93x slower     268.44 MOPS     3.15 MB
max                             4096x 4096    1.52 ms (   11.02 GFLOPS    44.10 GB/s) in torch,    4.58 ms (    3.67 GFLOPS    14.67 GB/s) in tinygrad,    3.01x slower      16.78 MOPS    67.11 MB
mul_sum                         4096x 4096    8.55 ms (    3.93 GFLOPS    15.70 GB/s) in torch,    8.66 ms (    3.88 GFLOPS    15.50 GB/s) in tinygrad,    1.01x slower      33.55 MOPS   134.22 MB
neg                             4096x 4096    5.89 ms (    2.85 GFLOPS    22.79 GB/s) in torch,    2.78 ms (    6.04 GFLOPS    48.31 GB/s) in tinygrad,    0.47x faster      16.78 MOPS   134.22 MB
conv bs:  1 chans: 12 ->  32 k:3              1.31 ms (   43.18 GFLOPS     1.11 GB/s) in torch,    1.18 ms (   48.06 GFLOPS     1.24 GB/s) in tinygrad,    0.90x faster      56.62 MOPS     1.46 MB
partial_sum                     4096x 4096    1.35 ms (   12.45 GFLOPS    49.78 GB/s) in torch,    2.15 ms (    7.81 GFLOPS    31.25 GB/s) in tinygrad,    1.59x slower      16.78 MOPS    67.11 MB
permute                         1024x 1024    1.21 ms (    0.87 GFLOPS     6.93 GB/s) in torch,    0.60 ms (    1.74 GFLOPS    13.95 GB/s) in tinygrad,    0.50x faster       1.05 MOPS     8.39 MB
permute                         4096x 4096   22.94 ms (    0.73 GFLOPS     5.85 GB/s) in torch,    2.81 ms (    5.98 GFLOPS    47.85 GB/s) in tinygrad,    0.12x faster      16.78 MOPS   134.22 MB
pow                             2048x 2048   20.83 ms (    4.83 GFLOPS     2.42 GB/s) in torch,    1.40 ms (   71.66 GFLOPS    35.83 GB/s) in tinygrad,    0.07x faster     100.66 MOPS    50.33 MB
relu                            4096x 4096    5.64 ms (    2.97 GFLOPS    23.79 GB/s) in torch,    2.78 ms (    6.04 GFLOPS    48.34 GB/s) in tinygrad,    0.49x faster      16.78 MOPS   134.22 MB
sub                             4096x 4096    6.71 ms (    2.50 GFLOPS    30.02 GB/s) in torch,    3.92 ms (    4.29 GFLOPS    51.42 GB/s) in tinygrad,    0.58x faster      16.78 MOPS   201.33 MB
sum                             2048x 2048    0.50 ms (    8.34 GFLOPS    33.36 GB/s) in torch,    1.25 ms (    3.36 GFLOPS    13.43 GB/s) in tinygrad,    2.48x slower       4.19 MOPS    16.78 MB
sum                             4096x 4096    1.33 ms (   12.61 GFLOPS    50.44 GB/s) in torch,    4.57 ms (    3.67 GFLOPS    14.69 GB/s) in tinygrad,    3.43x slower      16.78 MOPS    67.11 MB
.
----------------------------------------------------------------------
Ran 31 tests in 6.650s

OK (skipped=8)
```